tests: Rename linearizability tests to robustness
