+++ .github/workflows/robustness-template.yaml
@@ -1,4 +1,4 @@
-name: Reusable Linearizability Workflow
+name: Reusable Robustness Workflow
 on:
   workflow_call:
     inputs:
@@ -43,12 +43,12 @@ jobs:
             make build
             ;;
           esac
-    - name: test-linearizability
+    - name: test-robustness
       run: |
         # Use --failfast to avoid overriding report generated by failed test
-        EXPECT_DEBUG=true GO_TEST_FLAGS='-v --count ${{ inputs.count }} --timeout ${{ inputs.testTimeout }} --failfast --run TestLinearizability' RESULTS_DIR=/tmp/linearizability make test-linearizability
+        EXPECT_DEBUG=true GO_TEST_FLAGS='-v --count ${{ inputs.count }} --timeout ${{ inputs.testTimeout }} --failfast --run TestRobustness' RESULTS_DIR=/tmp/results make test-robustness
     - uses: actions/upload-artifact@v2
       if: always()
       with:
         name: ${{ inputs.artifactName }}
-        path: /tmp/linearizability/*
+        path: /tmp/results/*
