+++ Makefile
@@ -33,9 +33,9 @@ test-e2e: build
 test-e2e-release: build
 	PASSES="release e2e" ./scripts/test.sh $(GO_TEST_FLAGS)
 
-.PHONY: test-linearizability
-test-linearizability:
-	PASSES="linearizability" ./scripts/test.sh $(GO_TEST_FLAGS)
+.PHONY: test-robustness
+test-robustness:
+	PASSES="robustness" ./scripts/test.sh $(GO_TEST_FLAGS)
 
 .PHONY: fuzz
 fuzz:
